requirements.txt
weaviate_client==3.2.1
openpyxl==3.0.5
PyYAML==5.4.1
